+++ haoskiosk/run.sh
@@ -2,13 +2,13 @@
 # shellcheck shell=bash
 # Clean up on exit:
 TTY0_DELETED="" #Need to set to empty string since runs with nounset=on (like set -u)
-trap '[ -n "$(jobs -p)" ] && kill $(jobs -p); [ -n "$TTY0_DELETED" ] && mknod -m 620 /dev/tty0 c 4 0 && mount -o remount,ro /dev; exit' INT TERM EXIT
+trap '[ -n "$(jobs -p)" ] && kill $(jobs -p); [ -n "$TTY0_DELETED" ] && mknod -m 620 /dev/tty0 c 4 0; exit' INT TERM EXIT
 ################################################################################
 # Add-on: HAOS Kiosk Display (haoskiosk)
 # File: run.sh
-# Version: 0.9.8
+# Version: 0.9.9
 # Copyright Jeff Kosowsky
-# Date: June 2025
+# Date: July 2025
 #
 #  Code does the following:
 #     - Import and sanity-check the following variables from HA/config.yaml
@@ -23,6 +23,7 @@ trap '[ -n "$(jobs -p)" ] && kill $(jobs -p); [ -n "$TTY0_DELETED" ] && mknod -m
 #         BROWSER_REFRESH
 #         SCREEN_TIMEOUT
 #         HDMI_PORT
+#         DEBUG_MODE
 #     - Hack to delete (and later restore) /dev/tty0 (needed for X to start)
 #     - Start X window system
 #     - Start Openbox window manager
@@ -32,54 +33,52 @@ trap '[ -n "$(jobs -p)" ] && kill $(jobs -p); [ -n "$TTY0_DELETED" ] && mknod -m
 ################################################################################
 bashio::log.info "Starting haoskiosk..."
 ### Get config variables from HA add-on & set environment variables
-HA_USERNAME=$(bashio::config 'ha_username')
-HA_USERNAME="${HA_USERNAME/null/}"
+function get_config () {
+    # First, use existing variable if already set (for debugging purposes)
+    # If not set, lookup configuration value
+    # If null, use optional second parameter or else ""
+    local VAR_NAME="$1"
+    local DEFAULT="${2:-}"
+    local MASK="${3:-}"
 
-HA_PASSWORD=$(bashio::config 'ha_password')
-HA_PASSWORD="${HA_PASSWORD/null/}"
+    local VALUE="${!VAR_NAME:-}" #Existing value for variable if set (note need default since running as 'nounset')
+    if [ -z "$VALUE" ]; then
+        VALUE=$(bashio::config "${VAR_NAME,,}")
+    fi
+    if [ "$VALUE" = "null" ] || [ "$VALUE" = "" ]; then
+        VALUE="$DEFAULT"
+    fi
 
-HA_URL=$(bashio::config 'ha_url')
-HA_URL="${HA_URL/null/}"
+    export "$VAR_NAME=$VALUE"
 
-HA_DASHBOARD=$(bashio::config 'ha_dashboard')
-HA_DASHBOARD="${HA_DASHBOARD/null/}"
+    if [ -z "$MASK" ]; then
+        bashio::log.info "$VAR_NAME=$VALUE"
+    else
+        bashio::log.info "$VAR_NAME=XXXXXX"
+    fi
+}
 
-HA_THEME=$(bashio::config 'ha_theme')
-HA_THEME="${HA_THEME/null/}"
-
-HA_SIDEBAR=$(bashio::config 'ha_sidebar')
-HA_SIDEBAR="${HA_SIDEBAR/null/}"
-
-LOGIN_DELAY=$(bashio::config 'login_delay')
-LOGIN_DELAY="${LOGIN_DELAY/null/}"
-
-ZOOM_LEVEL=$(bashio::config 'zoom_level')
-ZOOM_LEVEL="${ZOOM_LEVEL/null/}"
-
-BROWSER_REFRESH=$(bashio::config 'browser_refresh')
-BROWSER_REFRESH="${BROWSER_REFRESH/null/}"
-
-export HA_USERNAME HA_PASSWORD HA_URL HA_DASHBOARD HA_THEME HA_SIDEBAR LOGIN_DELAY ZOOM_LEVEL BROWSER_REFRESH #Referenced in 'userconfig.lua'
-
-SCREEN_TIMEOUT=$(bashio::config 'screen_timeout')
-SCREEN_TIMEOUT="${SCREEN_TIMEOUT/null/}"
-SCREEN_TIMEOUT="${SCREEN_TIMEOUT:-600}" #Default to 600 seconds
-
-HDMI_PORT=$(bashio::config 'hdmi_port')
-HDMI_PORT="${HDMI_PORT/null/}"
-HDMI_PORT="${HDMI_PORT:-0}"
+get_config HA_USERNAME
+get_config HA_PASSWORD "" 1 #Don't log password
+get_config HA_URL
+get_config HA_DASHBOARD
+get_config HA_THEME
+get_config HA_SIDEBAR
+get_config LOGIN_DELAY
+get_config ZOOM_LEVEL
+get_config BROWSER_REFRESH
+get_config SCREEN_TIMEOUT 600 # Default to 600 seconds
+get_config HDMI_PORT 0 # Default to 0
 #NOTE: For now, both HDMI ports are mirrored and there is only /dev/fb0
 #      Not sure how to get them unmirrored so that console can be on /dev/fb0 and X on /dev/fb1
 #      As a result, setting HDMI=0 vs. 1 has no effect
+get_config DEBUG_MODE false
 
 #Validate environment variables set by config.yaml
 if [ -z "$HA_USERNAME" ] || [ -z "$HA_PASSWORD" ]; then
     bashio::log.error "Error: HA_USERNAME and HA_PASSWORD must be set"
     exit 1
 fi
-
-bashio::log.info "HA_USERNAME=$HA_USERNAME HA_PASSWORD=XXXXX HA_URL=$HA_URL HA_DASHBOARD=$HA_DASHBOARD HA_THEME=$HA_THEME HA_SIDEBAR=$HA_SIDEBAR"
-bashio::log.info "LOGIN_DELAY=$LOGIN_DELAY ZOOM_LEVEL=$ZOOM_LEVEL BROWSER_REFRESH=$BROWSER_REFRESH SCREEN_TIMEOUT=$SCREEN_TIMEOUT HDMI_PORT=$HDMI_PORT"
 
 ################################################################################
 ### Avoid waiting for DBUS timeouts (e.g., luakit)
@@ -92,6 +91,8 @@ rm -rf /tmp/.X*-lock #Cleanup old versions
 #because X doesn't have permissions to access it in the container
 #First, remount /dev as read-write since X absolutely, must have /dev/tty access
 #Note: need to use the version in util-linux, not busybox
+#Note: Do *not* later remount as 'ro' since that affect the root fs and
+#      in particular will block HAOS updates
 if [ -e "/dev/tty0" ]; then
     bashio::log.info "Attempting to (temporarily) delete /dev/tty0..."
     mount -o remount,rw /dev
@@ -99,8 +100,7 @@ if [ -e "/dev/tty0" ]; then
         bashio::log.error "Failed to remount /dev as read-write..."
         exit 1
     fi
-    if  ! rm /dev/tty0 ; then
-        mount -o remount,ro /dev
+    if  ! rm -f /dev/tty0 ; then
         bashio::log.error "Failed to delete /dev/tty0..."
         exit 1
     fi
@@ -118,12 +118,12 @@ for ((i=0; i<=XSTARTUP; i++)); do
     sleep 1
 done
 
-#Restore /dev/tty0 and 'ro' mode for /dev if deleted
+#Restore /dev/tty0
 if [ -n "$TTY0_DELETED" ]; then
-    if ( mknod -m 620 /dev/tty0 c 4 0 &&  mount -o remount,ro /dev ); then
+    if mknod -m 620 /dev/tty0 c 4 0; then
         bashio::log.info "Restored /dev/tty0 successfully..."
     else
-        bashio::log.error "Failed to restore /dev/tty0 and remount /dev/ read only..."
+        bashio::log.error "Failed to restore /dev/tty0..."
     fi
 fi
 
@@ -172,8 +172,12 @@ fi
     done
 )&
 
-#luakit -U "$HA_URL/$HA_DASHBOARD" #DEBUG
-#exec sleep 100000 #DEBUG
-### Run Luakit in the foreground
-bashio::log.info "Launching Luakit browser..."
-exec luakit -U "$HA_URL/$HA_DASHBOARD"
+echo "DEBUG_MODE=$DEBUG_MODE" #JJKCRAP
+if [ "$DEBUG_MODE" != true ]; then
+    ### Run Luakit in the foreground
+    bashio::log.info "Launching Luakit browser..."
+    exec luakit -U "$HA_URL/$HA_DASHBOARD"
+else ### Debug mode
+    bashio::log.info "Entering debug mode (X & Openbox but no luakit browser)..."
+    exec sleep infinite
+fi