+++ haoskiosk/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## v0.9.9 - July 2025
+
+- Removed remounting of /dev/ ro (which caused HAOS updates to fail)
+- Added 'debug' config that stops add-on before launching luakit
+- Cleaned up/improved code in run.sh and userconf.lua
+- Reverted to luakit=2.3.6-r0 since luakit=2.4.0-r0 crashes (temporary fix)
+
 ## v0.9.8 – June 2025
 
 - Added ability to set browser theme and sidebar behavior
@@ -14,7 +21,7 @@
 ## v0.9.7 – April 2025
 
 - Initial public release
-- Added Zoom capabilit
+- Added Zoom capability
 
 ## 0.9.6 – March 2025